ref(sidepanel): use modal and remove duplicated code
