+++ packages/ng/sidepanel/src/lib/index.ts
@@ -1,10 +1,3 @@
-export * from './sidepanel-config.default';
-export * from './sidepanel-config.model';
-export * from './sidepanel-panel.component';
-export * from './sidepanel-ref.factory';
-export * from './sidepanel-ref.model';
 export * from './sidepanel.model';
 export * from './sidepanel.module';
 export * from './sidepanel.service';
-export * from './sidepanel.token';
-export * from './sidepanel.translate';
