--- packages/ng/sidepanel/src/lib/sidepanel-config.model.ts
@@ -1,3 +1,0 @@
-import { LuModalConfig } from '@lucca-front/ng/modal';
-
-export type LuSidepanelConfig = LuModalConfig;
