+++ packages/ng/sidepanel/src/lib/sidepanel.model.ts
@@ -1,3 +1,56 @@
-import { ILuModalContent } from '@lucca-front/ng/modal';
+import { InjectionToken } from '@angular/core';
+import { ILuTranslation } from '@lucca-front/ng/core';
+import { ILuModalContent, ILuModalLabel, ILuModalRef, luDefaultModalConfig, LuModalConfig, luModalTranslations, LU_MODAL_DATA } from '@lucca-front/ng/modal';
 
+/**
+ * For backward compatibility, we re-export modal tokens as sidepanel tokens.
+ * We should remove all sidepanel related files in the next major version (16+).
+ */
+
+/**
+ * @deprecated Use LU_MODAL_CONFIG from @lucca-front/ng/modal instead.
+ */
+export const LU_SIDEPANEL_CONFIG = new InjectionToken<LuSidepanelConfig>('LuSidepanelDefaultConfig', {
+	factory: () => ({
+		...luDefaultModalConfig,
+		mode: 'sidepanel',
+		position: 'right',
+	}),
+});
+
+/**
+ * @deprecated Use LU_MODAL_DATA from @lucca-front/ng/modal instead.
+ */
+export const LU_SIDEPANEL_DATA = LU_MODAL_DATA;
+
+/**
+ * @deprecated Use LU_MODAL_TRANSLATIONS from @lucca-front/ng/modal instead.
+ */
+export const LU_SIDEPANEL_TRANSLATIONS = new InjectionToken('LuSidepanelTranslations', {
+	factory: () => luSidepanelTranslations,
+});
+
+/**
+ * @deprecated Use ILuModalLabel from @lucca-front/ng/modal instead.
+ */
+export type ILuSidepanelLabel = ILuModalLabel;
+
+/**
+ * @deprecated Use luModalTranslations from @lucca-front/ng/modal instead.
+ */
+export const luSidepanelTranslations: ILuTranslation<ILuSidepanelLabel> = luModalTranslations;
+
+/**
+ * @deprecated Use LuModalConfig from @lucca-front/ng/modal instead.
+ */
+export type LuSidepanelConfig = LuModalConfig;
+
+/**
+ * @deprecated Use ILuModalContent from @lucca-front/ng/modal instead.
+ */
 export type ILuSidepanelContent<T = unknown> = ILuModalContent<T>;
+
+/**
+ * @deprecated Use ILuModalRef from @lucca-front/ng/modal instead.
+ */
+export type ILuSidepanelRef<T extends ILuSidepanelContent = ILuSidepanelContent, D = unknown, R = unknown> = ILuModalRef<T, D, R>;
